+++ service/collector/error.go
@@ -62,3 +62,22 @@ var missingOrganizationLabel = &microerror.Error{
 func IsMissingOrganizationLabel(err error) bool {
 	return microerror.Cause(err) == missingOrganizationLabel
 }
+
+func IsNotFound(err error) bool {
+	if err == nil {
+		return false
+	}
+
+	c := microerror.Cause(err)
+
+	{
+		dErr, ok := c.(autorest.DetailedError)
+		if ok {
+			if dErr.StatusCode == 404 {
+				return true
+			}
+		}
+	}
+
+	return false
+}